extract duplicated service into module
